refactor front + models

+++ MedicalCentre/Models/StorageItem.cs
@@ -5,11 +5,11 @@ public class StorageItem
     public uint Id { get; set; }
     public string Name { get; set; }
     public uint Amount { get; set; }
-    public decimal Cost { get; set; }
+    public uint Cost { get; set; }
 
     public StorageItem() { }
 
-    public StorageItem(uint id, string name, uint amount, decimal cost)
+    public StorageItem(uint id, string name, uint amount, uint cost)
     {
         Id = id;
         Name = name;
